first draft of Event parsing
The Event struct is similar to header::Header, but all fields
are Option<T>. Only the fields that are valid event infomation are kept.

This first implementation keeps the fields corresponding to marker and
antenna position, in order to parse AntenaBeingMoved and
NewSiteOccupation events.

A new() fn is implemented to build an Event from the corresponding RINEX
section, and the Display trait is implemented for writing events.

A basic test that parses and displays a block checking the output is the
same as the input

Some errors are implemented for wrong block lengths, parsing of floats,
and for not yet implemented cases.