Stop ignoring events. Reorder Flags
Changed implementation of ord and PartialOrd for EpochFlags, in order
for events to ve printed before observations (for the same epoch).

Simple implementation for parse_event that returns an empty observation
after checking the event length is rigth.

The event information cannot be returned from the observation parsing,
because it cannot be represented as
BTreeMap<SV,HashMap<Observable,ObservationData>>

The task is left for the main parser, that should detect the Flag and
call the parser in event::

all previous tests passing without problem.