event::Record type and event::mixed::Record type
Create a BTreeMap type for storing spetial Event data of observations
files.

Al split, merge, etc, functionality should be implemented on this type.

To format into a file, we need to merge observation and event records,
for this i created the mixed::Record type, that is different from other
record types and is only meant for output.

Also in this commit, a formatter for the epoch of the event that
includes the number of lines that the event spans.