a new Record instance for observation with events.
ObsEvtRecord is different from the rest of the members in that
it encapsulates two different records, an observation::Record and an
event::Record.

The parsing function for observation data was modified to store event
data in a event::Record when parsing. At the end of the parsing run, if
the event record is not empty,  a ObsEvtRecord is returned. Otherwise
the behavior is unaltered.

Displaying ObsEvtRecord implies conversion of the underlying
records into a mixed::Record in order to produce a file with the right
ordering. All other functionalities should be implemented on the event::Record
level.